requirements.txt
PyQT5
tensorflow
numpy
pillow